The refactor of my entity is my enemy
Refactored some code!
~Player is now a type of Entity
~Made everything a little neater and we now have very few public fields
which is always good
~Changed the way the collision layer works. Now instead of having
certain terrain be "blocked" we just apply a collision map over the
whole layer which is nice in case we need specific cases for blocked or
unblocked tiles that look the same.
~ Fixed some collision stuff
+Added BlobEnemy a new type of Entity
- Removed old tile set (well it's still in the files but no longer in
code
+ Added a new Tileset

Probably forgetting a lot of stuff but this was a huge change so please
forgive me I've been refactoring all day