Dynamic level loading
[+] Implemented a form of dynamic level loading wherein while loading
the map the game searches for certain tags on the different tiles. If it
finds a tag it's looking for it puts an object in that tile (so far only
"start" and "blob" are functional, but it will be trivial to add further
objects)
[+] Blobs can actually die if they lose all their health, however the
player still does not have an attack to capitalize off of this fact
[~] Bounding boxes are off from the player and blob sprites. I need to
find out how to fix this just to soothe my headache. Since they're all
off by the same amount it actually doesn't matter for collision between
player and enemies

Notes:
- I did a sort of stress test on the game and learned that on PC it's
best not to go above 10k entities as at 10k entities we get around 35
fps which is not ideal (obviously)
- on android that 35 fps mark is around 300 entities (Samsung Galaxy S5
Neo)
- at 10k entities the app actually crashes because it runs out of RAM
and it runs at literally 1 frame per second
- interesting to note at a certain fps it seems like you're unable to
register a longpress because of its short duration perhaps?